Upgrade to Laravel 12 (#1006)
* Upgrade to Laravel 12

* Fix rector and phpstan issues

* Fix remaining linting issues

* Disable CarbonToDateFacadeRector rule

* Replace Date with Carbon for date manipulation

* Fix phpcs issues

* Fix larastan issues

* Refactor code

* Fix Carbon import

* Fully revert `Carbon` change

* Remove Laravel 12 ToDo

---------

+++ src/backend/app/Models/Person/Person.php
@@ -206,7 +206,7 @@ class Person extends BaseModel implements HasAddressesInterface, RoleInterface {
             return false;
         }
 
-        return Carbon::parse($lastPayment->created_at)->diffInDays(now()) <= 25;
+        return (int) Carbon::parse($lastPayment->created_at)->diffInDays(now()) <= 25;
     }
 
     /**
